+++ Front/my-react-app/src/components/objectOfLaborMaterial/objectOfLaborMaterialList.jsx
@@ -1,10 +1,12 @@
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import api from "../../api/api";
 import "../style/style.css"
 import ObjectOfLaborMaterial from "./objectOfLaborMaterial";
 import AddObjectOfLaborMaterial from "./addObjectOfLaborMaterial";
+import AuthContext from "../../context/authContext";
+import swal from "sweetalert";
 const ObjectOfLaborMaterialList=({objectOfLaborId})=>{
-
+const context=useContext(AuthContext);
 const [objectOfLaborMaterials, setObjectOfLaborMaterials]=useState();
 const [isAddModalOpen, setAddModalOpen] = useState(false);
 
@@ -18,9 +20,16 @@ const [isAddModalOpen, setAddModalOpen] = useState(false);
   };
 
   const handleAdd = async (newData) => {
+    if(context.type()!="Admin")
+    {
+      alert("Dodavanje je dozvoljeno samo administratoru.");
+      return;
+    }
     try {
       const res = await api.post("objectOfLaborMaterial", newData);
-        fetch();       
+      swal("Uspesno ste dodali novi entitet!", "", "success")
+  
+      fetch();       
     } catch (error) {
       alert(error);
     }
@@ -44,6 +53,11 @@ useEffect(() => {
   
 
   const handleDelete = async(id) => {
+    if(context.type()!="Admin")
+    {
+      alert("Brisanje je dozvoljeno samo administratoru.");
+      return;
+    }
     try {
         await api.delete('objectOfLaborMaterial/' + id);
             fetch();
