+++ Front/my-react-app/src/components/organizationalUnit/organizationalUnitList.jsx
@@ -1,11 +1,12 @@
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import api from "../../api/api";
 import "../style/style.css"
 import OrganizationalUnit from "./otganiaztionalUnit";
 import AddOrganizationalUnit from "./addOrganiaztionalUnit";
-
+import AuthContext from "../../context/authContext";
+import swal from "sweetalert";
 const OrganizationalUnitList=()=>{
-
+const context=useContext(AuthContext);
 const [organizationalUnits, setOrganizationalUnits]=useState();
 const [isAddModalOpen, setAddModalOpen] = useState(false);
 
@@ -19,9 +20,16 @@ const [isAddModalOpen, setAddModalOpen] = useState(false);
   };
 
   const handleAdd = async (newData) => {
+    if(context.type()!="Admin")
+    {
+      alert("Dodavanje je dozvoljeno samo administratoru.");
+      return;
+    }
     try {
       const res = await api.post("organizationalUnit", newData);
-        fetch();
+      swal("Uspesno ste dodali novi entitet!", "", "success")
+  
+      fetch();
     } catch (error) {
       alert(error);
     }
@@ -46,7 +54,11 @@ useEffect(() => {
   };
   
   const handleDelete = async(id) => {
-    console.log(id);
+    if(context.type()!="Admin")
+    {
+      alert("Brisanje je dozvoljeno samo administratoru.");
+      return;
+    }
     try {
         await api.delete('organizationalUnit/' + id);
             fetch();
