+++ Front/my-react-app/src/components/productionOrder/productionOrder.jsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import api from "../../api/api";
 import UpdateProductionOrder from "./updateProductionOrder";
+import swal from "sweetalert";
 const ProductionOrder=({data, onDelete, update})=>{
     const [productionOrder, setProductionOrder] = useState({});
     const [showModal, setShowModal]=useState(false);
@@ -13,10 +14,21 @@ const ProductionOrder=({data, onDelete, update})=>{
 
 
 
-   const handleDelete=(e)=>{
+   const handleDelete=async(e)=>{
 
     e.preventDefault();
-    onDelete();
+    const willDelete = await swal({
+      title: "Da li ste sigurni?",
+      text: "Da li ste sigurni da zelite da obrisete obrisete ovaj entitet?",
+      icon: "warning",
+      dangerMode: true,
+      buttons: ["Ne", true]
+    });
+    
+    if (willDelete) {
+      swal("Obrisano!", "", "success");
+      onDelete(); 
+    }
    };
