+++ Front/my-react-app/src/components/productionPlan/productionPlan.jsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import api from "../../api/api";
 import UpdateProductionPlan from "./updateProductionPlan";
+import swal from "sweetalert";
 const ProductionPlan=({data, onDelete, update})=>{
 
     const [productionPlan, setProductionPlan] = useState({});
@@ -15,10 +16,21 @@ const ProductionPlan=({data, onDelete, update})=>{
 
 
 
-    const handleDelete=(e)=>{
+    const handleDelete=async(e)=>{
 
       e.preventDefault();
-      onDelete();
+      const willDelete = await swal({
+        title: "Da li ste sigurni?",
+        text: "Da li ste sigurni da zelite da obrisete obrisete ovaj entitet?",
+        icon: "warning",
+        dangerMode: true,
+        buttons: ["Ne", true]
+      });
+      
+      if (willDelete) {
+        swal("Obrisano!", "", "success");
+        onDelete(); 
+      }
      };
