+++ Front/my-react-app/src/components/warehouse/warehouseList.jsx
@@ -1,14 +1,16 @@
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import api from "../../api/api";
 import "../style/style.css"
 import Warehouse from "./warehouse";
 import AddWarehouse from "./addWarehouse";
+import swal from "sweetalert";
+import AuthContext from "../../context/authContext";
 
 const WarehouseList=()=>{
 
 const [warehouses, setWarehouses]=useState();
 const [isAddModalOpen, setAddModalOpen] = useState(false);
-
+const context=useContext(AuthContext);
   const handleOpenAddModal = () => {
     setAddModalOpen(true);
   };
@@ -19,9 +21,16 @@ const [isAddModalOpen, setAddModalOpen] = useState(false);
   };
 
   const handleAdd = async (newData) => {
-    console.log("Dodavanje skladiste:", newData);
+    if(context.type()!="Admin")
+    {
+      alert("Dodavanje je dozvoljeno samo administratoru.");
+      return;
+    }    
+    
     try {
       const res = await api.post("warehouse", newData);
+      swal("Uspesno ste dodali novi entitet!", "", "success")
+
       fetch();
     } catch (error) {
       alert(error);
@@ -47,6 +56,11 @@ useEffect(() => {
   };
   
   const handleDelete = async(id) => {
+    if(context.type()!="Admin")
+    {
+      alert("Brisanje je dozvoljeno samo administratoru.");
+      return;
+    }
     try {
         await api.delete('warehouse/' + id);
             fetch();
